fix(test): address review — deduplicate, fix assertions, shared helpers
- Merge FullProjectSearchPerformanceTests into ProjectSearchPerformanceTests
- Merge LargeFileOpenPerformanceTests into EditorStressTests
- Move XCTAssertEqual/XCTAssertGreaterThan outside measure {} blocks
- Extract generateCode() into shared PerformanceTestHelpers enum
- Extract createTextSystem() into XCTestCase extension with XCTFail + throw
- Move ensureLayout(for:) before measure {} for realistic benchmarking
- Remove unused `_ = lineRect.origin.y * 0.12` expression
- Replace python3 with jq in nightly workflow summary
- Increase retention-days from 14 to 30
- Add GitHub issue creation step on nightly test failure
- Add issues: write permission to nightly workflow
