+++ PinePerformanceTests/MinimapRenderPerformanceTests.swift
@@ -35,48 +35,21 @@ final class MinimapRenderPerformanceTests: XCTestCase {
     // MARK: - Helpers
 
     private func generateCode(lines: Int) -> String {
-        var result: [String] = ["import Foundation", ""]
-        var classIdx = 0
-        var lineCount = 2
-        while lineCount < lines {
-            result.append("class Minimap\(classIdx) {")
-            result.append("    var value: Int = \(classIdx)")
-            for method in 0..<4 {
-                guard lineCount + 6 < lines else { break }
-                result.append("    func render\(method)() -> String {")
-                result.append("        let x = value * \(method)")
-                result.append("        return \"pixel: \\(x)\"")
-                result.append("    }")
-                lineCount += 4
-            }
-            result.append("}")
-            result.append("")
-            lineCount += 4
-            classIdx += 1
-        }
-        return result.joined(separator: "\n")
+        PerformanceTestHelpers.generateSwiftCode(lines: lines, classPrefix: "Minimap")
     }
 
     /// Creates a fully configured text system with highlighted text
-    /// and returns (textView, textStorage) pair.
-    private func createTextSystem(lines: Int) -> (NSTextView, NSTextStorage) {
+    /// and forced layout. Throws on failure instead of returning empty objects.
+    private func createHighlightedTextSystem(
+        lines: Int
+    ) throws -> (NSTextView, NSTextStorage) {
         let code = generateCode(lines: lines)
-
-        // Create NSTextView first, then use its built-in text system
-        let textView = NSTextView(frame: NSRect(x: 0, y: 0, width: 800, height: 100_000))
-        textView.textContainer?.size = NSSize(width: 800, height: CGFloat.greatestFiniteMagnitude)
-        textView.textContainer?.widthTracksTextView = true
-
-        guard let textStorage = textView.textStorage else {
-            return (textView, NSTextStorage())
-        }
-
-        textStorage.setAttributedString(NSAttributedString(string: code))
+        let (textView, textStorage, _) = try createTextSystem(text: code)
 
         let font = NSFont.monospacedSystemFont(ofSize: 12, weight: .regular)
         highlighter.highlight(textStorage: textStorage, language: "minimapperfswift", font: font)
 
-        // Force layout so enumerateLineFragments works
+        // Force layout so enumerateLineFragments works — done before measure
         if let layoutManager = textView.layoutManager, let textContainer = textView.textContainer {
             layoutManager.ensureLayout(for: textContainer)
         }
@@ -89,8 +62,8 @@ final class MinimapRenderPerformanceTests: XCTestCase {
     /// Benchmarks the data preparation phase for minimap rendering:
     /// walking line fragments and extracting syntax color segments.
     /// This is the CPU-intensive part of MinimapView.draw().
-    func testMinimapDataPrep2000Lines() {
-        let (textView, textStorage) = createTextSystem(lines: 2000)
+    func testMinimapDataPrep2000Lines() throws {
+        let (textView, textStorage) = try createHighlightedTextSystem(lines: 2000)
 
         guard let layoutManager = textView.layoutManager,
               let textContainer = textView.textContainer else {
@@ -135,8 +108,8 @@ final class MinimapRenderPerformanceTests: XCTestCase {
     }
 
     /// Benchmarks minimap data preparation for a 5000-line file.
-    func testMinimapDataPrep5000Lines() {
-        let (textView, textStorage) = createTextSystem(lines: 5000)
+    func testMinimapDataPrep5000Lines() throws {
+        let (textView, textStorage) = try createHighlightedTextSystem(lines: 5000)
 
         guard let layoutManager = textView.layoutManager,
               let textContainer = textView.textContainer else {
@@ -172,8 +145,8 @@ final class MinimapRenderPerformanceTests: XCTestCase {
     /// Benchmarks minimap data preparation with diff markers overlaid.
     /// Simulates the real minimap render path: walk line fragments,
     /// extract syntax colors, and check each line against a diff map.
-    func testMinimapDataPrepWithDiffMarkers() {
-        let (textView, textStorage) = createTextSystem(lines: 3000)
+    func testMinimapDataPrepWithDiffMarkers() throws {
+        let (textView, textStorage) = try createHighlightedTextSystem(lines: 3000)
 
         guard let layoutManager = textView.layoutManager,
               let textContainer = textView.textContainer else {
@@ -193,12 +166,10 @@ final class MinimapRenderPerformanceTests: XCTestCase {
             var lineIndex = 0
             var markerCount = 0
 
-            layoutManager.enumerateLineFragments(forGlyphRange: fullRange) { lineRect, _, _, glyphRange, _ in
+            layoutManager.enumerateLineFragments(forGlyphRange: fullRange) { _, _, _, glyphRange, _ in
                 let charRange = layoutManager.characterRange(
                     forGlyphRange: glyphRange, actualGlyphRange: nil
                 )
-                _ = lineRect.origin.y * 0.12
-
                 // Syntax color segments (same as real minimap)
                 var pos = charRange.location
                 let end = NSMaxRange(charRange)
@@ -223,8 +194,8 @@ final class MinimapRenderPerformanceTests: XCTestCase {
     }
 
     /// Benchmarks the full minimap pipeline: highlight + data preparation.
-    func testMinimapFullPipeline3000Lines() {
-        let (textView, textStorage) = createTextSystem(lines: 3000)
+    func testMinimapFullPipeline3000Lines() throws {
+        let (textView, textStorage) = try createHighlightedTextSystem(lines: 3000)
         let font = NSFont.monospacedSystemFont(ofSize: 12, weight: .regular)
 
         guard let layoutManager = textView.layoutManager,
